+++ targets/neoforge-26.1/src/main/java/com/sighs/sophisticatedsorter/SophisticatedSorter.java
@@ -4,6 +4,7 @@ import com.mojang.logging.LogUtils;
 import com.sighs.sophisticatedsorter.settings.ContainerMemorySlotGuard;
 import com.sighs.sophisticatedsorter.settings.ContainerSettingsTracker;
 import com.sighs.sophisticatedsorter.settings.ModMenus;
+import com.sighs.sophisticatedsorter.common.CorePrerequisite;
 import com.sighs.sophisticatedsorter.utils.CoreUtils;
 import com.sighs.sophisticatedsorter.utils.PlatformSortBackend;
 import net.neoforged.api.distmarker.Dist;
@@ -24,6 +25,7 @@ public class SophisticatedSorter {
     public static final Logger LOGGER = LogUtils.getLogger();
 
     public SophisticatedSorter(IEventBus modEventBus, ModContainer modContainer) {
+        CorePrerequisite.requireCore();
         CoreUtils.installPlatform(PlatformSortBackend.INSTANCE);
         modContainer.registerConfig(ModConfig.Type.CLIENT, Config.SPEC);
         modContainer.registerConfig(ModConfig.Type.SERVER, Config.SERVER_SPEC);